==> scripts/00-packages.R <==
# All packages used (at some point)
devtools::install_github("KluaneRedSquirrelProject/krsp")
library(krsp)
library(data.table)
library(dplyr)
library(lubridate)
library(sf)
library(ggplot2)
library(ggmap)
library(hms)
library(sp)
library(tidyverse)
library(tidyr)
library(concaveman)
library(adehabitatHR)
library(amt)
library(raster)
library(arm)
library(rsq)
